ECL-496 - updated revenue from Long to BigDecimal

+++ app/uk/gov/hmrc/economiccrimelevyreturns/models/CalculatedLiability.scala
@@ -43,7 +43,7 @@ object Band {
   }
 }
 
-final case class EclAmount(amount: Double, apportioned: Boolean)
+final case class EclAmount(amount: BigDecimal, apportioned: Boolean)
 
 object EclAmount {
   implicit val format: OFormat[EclAmount] = Json.format[EclAmount]
